requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
pymysql==1.0.3
pydantic==1.10.7
pydantic[email]==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator<2.0,>=1.1
python-dotenv==1.0.0
fastapi-mail==1.2.8